move a couple useful functions into `mirage/utils`
pagination espeically; re using that on the search endpoint
